requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
iso8601==1.0.2
m3u8==1.0.0
pymediainfo==5.1.0
PySocks==1.7.1
requests==2.27.1
urllib3==1.26.9
